feat: support for common units (#420)
## Context
See #417

## Change Summary
- Autogenerate the common base units as suggested in the original issue
- Update union/optional syntax for 3.10
- Code cleanup removing some redundant things like `__ne__` when
`__eq__` is implemented
- Changed str and repr for Units, these were displaying internal
attributes and weren't super helpful. Generally, str should be a nice,
formatted version of an object in e.g. a quantity is {value} {units}.
Repr should be if something is easy to construct the string that can
reproduce the same value when evaled e.g. eval(repr(quantity)) ==
quantity
- 

## Rationale
This PR makes code much more concise and easier to understand. It also
modernises the codebase a bit by fixing some of the type checking
issues.

## Impact
Closes #417

---------

Co-authored-by: Harshal Pohekar <106588300+hpohekar@users.noreply.github.com>